_533-django-example-yard_try-rules/533-dkrs-djangoexample/mangosite/requirements.txt
Django>=2,<3
django-extensions
djangorestframework
djangorestframework_simplejwt
rules